@@ -27,11 +27,11 @@
     "@storybook/addon-actions": "^5.2.1",
     "@storybook/react": "^5.2.1",
     "@wdio/browserstack-service": "^5.13.2",
-    "@wdio/cli": "^5.13.2",
-    "@wdio/local-runner": "^5.13.2",
-    "@wdio/mocha-framework": "^5.13.2",
+    "@wdio/cli": "^5.14.3",
+    "@wdio/local-runner": "^5.14.3",
+    "@wdio/mocha-framework": "^5.14.3",
     "@wdio/spec-reporter": "^5.13.2",
-    "@wdio/sync": "^5.13.2",
+    "@wdio/sync": "^5.14.3",
     "aigle": "^1.14.1",
     "autobind-decorator": "^2.4.0",
     "autoprefixer": "^9.6.1",
@@ -43,7 +43,7 @@
     "chalk": "^2.4.2",
     "chromedriver": "^77.0.0",
     "colors": "^1.4.0",
-    "config": "^3.2.2",
+    "config": "^3.2.3",
     "copy-webpack-plugin": "^5.0.4",
     "cross-fetch": "^3.0.4",
     "cross-spawn": "^7.0.0",
@@ -66,7 +66,7 @@
     "file-loader": "^4.2.0",
     "git-describe": "^4.0.4",
     "graphql-cli": "^3.0.14",
-    "handlebars": "^4.3.1",
+    "handlebars": "^4.4.2",
     "handlebars-loader": "^1.7.1",
     "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.2.0",
@@ -78,22 +78,22 @@
     "postcss-flexbugs-fixes": "^4.1.0",
     "postcss-loader": "^3.0.0",
     "prettier": "^1.18.2",
-    "react-test-renderer": "^16.9.0",
+    "react-test-renderer": "^16.10.2",
     "selenium-webdriver": "^4.0.0-alpha.5",
     "style-loader": "^1.0.0",
     "threads": "^0.12.0",
     "uuid": "^3.3.3",
     "wdio-chromedriver-service": "^5.0.2",
-    "webdriverio": "^5.13.2",
+    "webdriverio": "^5.14.3",
     "webpack": "^4.41.0",
     "webpack-cleanup-plugin": "^0.5.1",
     "webpack-cli": "^3.3.9",
-    "webpack-dev-server": "^3.8.1",
+    "webpack-dev-server": "^3.8.2",
     "webpack-merge": "^4.2.2",
     "xhr2": "^0.2.0"
   },
   "dependencies": {
-    "@apollo/react-hooks": "^3.1.1",
+    "@apollo/react-hooks": "^3.1.2",
     "@blueprintjs/core": "^3.18.1",
     "@blueprintjs/datetime": "^3.9.0",
     "@fullcalendar/core": "^4.3.1",
@@ -131,18 +131,18 @@
     "moment": "^2.24.0",
     "pluralize": "^8.0.0",
     "prop-types": "^15.7.2",
-    "react": "^16.9.0",
+    "react": "^16.10.2",
     "react-avatar-edit": "^0.8.3",
     "react-bootstrap": "^0.32.4",
     "react-confirm-bootstrap": "^5.3.0",
     "react-container-dimensions": "^1.4.1",
-    "react-dom": "^16.9.0",
+    "react-dom": "^16.10.2",
     "react-mosaic-component": "^3.2.0",
     "react-notification-badge": "^1.5.1",
     "react-redux": "^7.1.1",
     "react-redux-loading-bar": "^4.4.0",
     "react-router-bootstrap": "^0.25.0",
-    "react-router-dom": "^5.0.1",
+    "react-router-dom": "^5.1.2",
     "react-scroll": "^1.7.14",
     "react-tag-input": "^5.2.3",
     "react-toastify": "^5.4.0",
