@@ -1,33 +1,31 @@
 import PropTypes from "prop-types"
-import React, { Component } from "react"
-import { Prompt, withRouter } from "react-router-dom"
+import React, { useEffect } from "react"
+import { Prompt } from "react-router-dom"
 
 const LEAVE_WARNING =
   "Are you sure you wish to navigate away from the page? You will lose unsaved changes."
 
-class NavigationWarning extends Component {
-  static propTypes = {
-    isBlocking: PropTypes.bool
-  }
+const NavigationWarning = props => {
+  const { isBlocking } = props
+  useEffect(() => {
+    window.addEventListener("beforeunload", onBeforeUnloadListener)
 
-  onBeforeUnloadListener = event => {
-    if (this.props.isBlocking) {
-      event.returnValue = LEAVE_WARNING
-      event.preventDefault()
+    return () =>
+      window.removeEventListener("beforeunload", onBeforeUnloadListener)
+
+    function onBeforeUnloadListener(event) {
+      if (isBlocking) {
+        event.returnValue = LEAVE_WARNING
+        event.preventDefault()
+      }
     }
-  }
+  }, [isBlocking])
 
-  componentDidMount() {
-    window.addEventListener("beforeunload", this.onBeforeUnloadListener)
-  }
-
-  componentWillUnmount() {
-    window.removeEventListener("beforeunload", this.onBeforeUnloadListener)
-  }
-
-  render() {
-    return <Prompt when={this.props.isBlocking} message={LEAVE_WARNING} />
-  }
+  return <Prompt when={isBlocking} message={LEAVE_WARNING} />
 }
 
-export default withRouter(NavigationWarning)
+NavigationWarning.propTypes = {
+  isBlocking: PropTypes.bool
+}
+
+export default NavigationWarning
