@@ -1,10 +1,9 @@
 import AppContext from "components/AppContext"
-import { routerRelatedPropTypes } from "components/Page"
 import * as Models from "models"
 import PropTypes from "prop-types"
-import React, { Component } from "react"
+import React from "react"
 import { Button, DropdownButton, MenuItem } from "react-bootstrap"
-import { withRouter } from "react-router-dom"
+import { useHistory } from "react-router-dom"
 
 const DEFAULT_ACTIONS = [Models.Report]
 
@@ -16,54 +15,52 @@ const ADMIN_ACTIONS = [
   Models.AuthorizationGroup
 ]
 
-class BaseCreateButton extends Component {
-  static propTypes = {
-    currentUser: PropTypes.instanceOf(Models.Person),
-    ...routerRelatedPropTypes
+const BaseCreateButton = props => {
+  const { currentUser } = props
+  const history = useHistory()
+
+  const modelClasses = DEFAULT_ACTIONS.concat(
+    currentUser.isSuperUser() && SUPER_USER_ACTIONS,
+    currentUser.isAdmin() && ADMIN_ACTIONS
+  ).filter(value => !!value)
+
+  if (modelClasses.length > 1) {
+    return (
+      <DropdownButton
+        title="Create"
+        pullRight
+        bsStyle="primary"
+        id="createButton"
+        onSelect={onSelect}
+      >
+        {modelClasses.map((modelClass, i) => (
+          <MenuItem key={modelClass.resourceName} eventKey={modelClass}>
+            New {modelClass.displayName() || modelClass.resourceName}
+          </MenuItem>
+        ))}
+      </DropdownButton>
+    )
+  } else if (modelClasses.length) {
+    let modelClass = modelClasses[0]
+    return (
+      <Button
+        bsStyle="primary"
+        onClick={() => onSelect(modelClass)}
+        id="createButton"
+      >
+        New{" "}
+        {(modelClass.displayName() || modelClass.resourceName).toLowerCase()}
+      </Button>
+    )
   }
 
-  render() {
-    const { currentUser } = this.props
-
-    const modelClasses = DEFAULT_ACTIONS.concat(
-      currentUser.isSuperUser() && SUPER_USER_ACTIONS,
-      currentUser.isAdmin() && ADMIN_ACTIONS
-    ).filter(value => !!value)
-
-    if (modelClasses.length > 1) {
-      return (
-        <DropdownButton
-          title="Create"
-          pullRight
-          bsStyle="primary"
-          id="createButton"
-          onSelect={this.onSelect}
-        >
-          {modelClasses.map((modelClass, i) => (
-            <MenuItem key={modelClass.resourceName} eventKey={modelClass}>
-              New {modelClass.displayName() || modelClass.resourceName}
-            </MenuItem>
-          ))}
-        </DropdownButton>
-      )
-    } else if (modelClasses.length) {
-      let modelClass = modelClasses[0]
-      return (
-        <Button
-          bsStyle="primary"
-          onClick={() => this.onSelect(modelClass)}
-          id="createButton"
-        >
-          New{" "}
-          {(modelClass.displayName() || modelClass.resourceName).toLowerCase()}
-        </Button>
-      )
-    }
+  function onSelect(modelClass) {
+    history.push(modelClass.pathForNew())
   }
+}
 
-  onSelect = modelClass => {
-    this.props.history.push(modelClass.pathForNew())
-  }
+BaseCreateButton.propTypes = {
+  currentUser: PropTypes.instanceOf(Models.Person)
 }
 
 const CreateButton = props => (
@@ -74,4 +71,4 @@ const CreateButton = props => (
   </AppContext.Consumer>
 )
 
-export default withRouter(CreateButton)
+export default CreateButton
