@@ -1,13 +1,13 @@
 import API, { Settings } from "api"
 import { gql } from "apollo-boost"
 import Calendar from "components/Calendar"
-import { mapDispatchToProps, routerRelatedPropTypes } from "components/Page"
+import { mapDispatchToProps } from "components/Page"
 import { Person, Report } from "models"
 import moment from "moment"
 import PropTypes from "prop-types"
 import React, { useRef } from "react"
 import { connect } from "react-redux"
-import { withRouter } from "react-router-dom"
+import { useHistory } from "react-router-dom"
 
 const GQL_GET_REPORT_LIST = gql`
   query($reportQuery: ReportSearchQueryInput) {
@@ -42,13 +42,14 @@ const GQL_GET_REPORT_LIST = gql`
 
 const ReportCalendar = props => {
   const { queryParams, setTotalCount } = props
+  const history = useHistory()
   const calendarComponentRef = useRef(null)
 
   return (
     <Calendar
       events={getEvents}
       eventClick={info => {
-        props.history.push(info.event.url)
+        history.push(info.event.url)
         // Prevent browser navigation to the url
         info.jsEvent.preventDefault()
       }}
@@ -98,11 +99,10 @@ const ReportCalendar = props => {
 
 ReportCalendar.propTypes = {
   queryParams: PropTypes.object,
-  setTotalCount: PropTypes.func,
-  ...routerRelatedPropTypes
+  setTotalCount: PropTypes.func
 }
 
 export default connect(
   null,
   mapDispatchToProps
-)(withRouter(ReportCalendar))
+)(ReportCalendar)
