@@ -3,6 +3,7 @@ import moment from "moment"
 export const BETWEEN = "0"
 export const BEFORE = "1"
 export const AFTER = "2"
+export const ON = "3"
 export const LAST_DAY = -1 * 1000 * 60 * 60 * 24
 export const LAST_WEEK = LAST_DAY * 7
 export const LAST_MONTH = LAST_DAY * 30
@@ -11,6 +12,7 @@ export const RANGE_TYPE_LABELS = {
   [BETWEEN]: "Between",
   [BEFORE]: "Before",
   [AFTER]: "After",
+  [ON]: "On",
   [LAST_DAY]: "Last 24 hours",
   [LAST_WEEK]: "Last 7 days",
   [LAST_MONTH]: "Last 30 days"
@@ -44,11 +46,17 @@ export function dateToQuery(queryKey, value) {
     return {
       [startKey]: moment(value.start).startOf("day")
     }
+  } else if (value.relative === ON) {
+    // On given date
+    return {
+      [startKey]: moment(value.start).startOf("day"),
+      [endKey]: moment(value.start).endOf("day")
+    }
   } else {
     // LAST_DAY, LAST_WEEK, LAST_MONTH => Time relative to now, up till now
     return {
       [startKey]: parseInt(value.relative),
-      [endKey]: moment()
+      [endKey]: moment().endOf("day")
     }
   }
 }
