@@ -12,6 +12,7 @@ import RelatedObjectNotes, {
 import { Location } from "models"
 import React from "react"
 import { connect } from "react-redux"
+import { useParams } from "react-router-dom"
 import LocationForm from "./Form"
 
 const GQL_GET_LOCATION = gql`
@@ -28,7 +29,7 @@ const GQL_GET_LOCATION = gql`
 `
 
 const LocationEdit = props => {
-  const uuid = props.match.params.uuid
+  const { uuid } = useParams()
   const { loading, error, data } = API.useApiQuery(GQL_GET_LOCATION, {
     uuid
   })
