@@ -12,6 +12,7 @@ import RelatedObjectNotes, {
 import { Position } from "models"
 import React from "react"
 import { connect } from "react-redux"
+import { useParams } from "react-router-dom"
 import PositionForm from "./Form"
 
 const GQL_GET_POSITION = gql`
@@ -58,7 +59,7 @@ const GQL_GET_POSITION = gql`
 `
 
 const PositionEdit = props => {
-  const uuid = props.match.params.uuid
+  const { uuid } = useParams()
   const { loading, error, data } = API.useApiQuery(GQL_GET_POSITION, {
     uuid
   })
