@@ -143,7 +143,9 @@ const AttendeesTable = props => {
           <LinkTo person={person} showIcon={false} />
         </td>
         <td>
-          <LinkTo position={person.position} />
+          {person.position && person.position.uuid && (
+            <LinkTo position={person.position} />
+          )}
           {person.position && person.position.code
             ? `, ${person.position.code}`
             : ""}
@@ -158,7 +160,7 @@ const AttendeesTable = props => {
           <LinkTo
             whenUnspecified=""
             organization={person.position && person.position.organization}
-          />{" "}
+          />
         </td>
         {showDelete && (
           <td>
