@@ -12,7 +12,7 @@ import RelatedObjectNotes, {
 import { Report } from "models"
 import React from "react"
 import { connect } from "react-redux"
-import { withRouter } from "react-router-dom"
+import { useParams } from "react-router-dom"
 import ReportForm from "./Form"
 
 const GQL_GET_REPORT = gql`
@@ -94,7 +94,7 @@ const GQL_GET_REPORT = gql`
 `
 
 const ReportEdit = props => {
-  const uuid = props.match.params.uuid
+  const { uuid } = useParams()
   const { loading, error, data } = API.useApiQuery(GQL_GET_REPORT, {
     uuid
   })
@@ -151,4 +151,4 @@ ReportEdit.propTypes = {
 export default connect(
   null,
   mapDispatchToProps
-)(withRouter(ReportEdit))
+)(ReportEdit)
