@@ -34,7 +34,7 @@ import React, { useMemo, useState } from "react"
 import { Button, HelpBlock, Modal } from "react-bootstrap"
 import ContainerDimensions from "react-container-dimensions"
 import { connect } from "react-redux"
-import { withRouter } from "react-router-dom"
+import { useHistory, useLocation } from "react-router-dom"
 import utils from "utils"
 
 const GQL_ROLLUP_GRAPH = gql`
@@ -94,9 +94,7 @@ const GQL_EMAIL_ROLLUP = gql`
       principalOrganizationUuid: $principalOrganizationUuid
       advisorOrganizationUuid: $advisorOrganizationUuid
       orgType: $orgType
-    ) {
-      uuid
-    }
+    )
   }
 `
 
@@ -256,7 +254,9 @@ Map.propTypes = {
 
 const BaseRollupShow = props => {
   const { appSettings, searchQuery } = props
-  const { startDate, endDate } = getDateRangeFromQS(props.location.search)
+  const history = useHistory()
+  const routerLocation = useLocation()
+  const { startDate, endDate } = getDateRangeFromQS(routerLocation.search)
   const [showEmailModal, setShowEmailModal] = useState(false)
   const [orgType, setOrgType] = useState(Organization.TYPE.ADVISOR_ORG)
   const [focusedOrg, setFocusedOrg] = useState(null)
@@ -486,7 +486,7 @@ const BaseRollupShow = props => {
     if (!startDate || !endDate) {
       return
     }
-    props.history.replace({
+    history.replace({
       pathname: "rollup",
       search: utils.formatQueryString({
         startDate,
@@ -658,4 +658,4 @@ const RollupShow = props => (
 export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(withRouter(RollupShow))
+)(RollupShow)
