@@ -12,6 +12,7 @@ import RelatedObjectNotes, {
 import { Task } from "models"
 import React from "react"
 import { connect } from "react-redux"
+import { useParams } from "react-router-dom"
 import TaskForm from "./Form"
 
 const GQL_GET_TASK = gql`
@@ -61,7 +62,7 @@ const GQL_GET_TASK = gql`
 `
 
 const TaskEdit = props => {
-  const uuid = props.match.params.uuid
+  const { uuid } = useParams()
   const { loading, error, data } = API.useApiQuery(GQL_GET_TASK, {
     uuid
   })
